[issue-1538] Update filter data by time and move logic from AppNavigator to AppOnlineContentProvider
