+++ src/AppNavigator.tsx
@@ -86,7 +86,6 @@ import { useGroupYieldPosition } from 'hooks/earning';
 import { AboutSubWallet } from 'screens/Settings/AboutSubWallet';
 import { updateCurrentRoute } from 'stores/utils';
 import { AppOnlineContentContext } from 'providers/AppOnlineContentProvider';
-import { GlobalModalContext } from 'providers/GlobalModalContext';
 
 interface Props {
   isAppReady: boolean;
@@ -212,23 +211,6 @@ type DeepLinkSubscriptionType = {
   url: string;
 };
 
-const getAppTransformRouteName = (currentRoute?: string) => {
-  if (!currentRoute) {
-    return '';
-  }
-
-  switch (currentRoute) {
-    case 'Tokens':
-      return 'token';
-    case 'NFTs':
-      return 'nft';
-    case 'Earning':
-      return 'earning';
-    case 'Crowdloans':
-      return 'crowdloan';
-  }
-};
-
 const AppNavigator = ({ isAppReady }: Props) => {
   const isDarkMode = true;
   const theme = isDarkMode ? THEME_PRESET.dark : THEME_PRESET.light;
@@ -253,9 +235,7 @@ const AppNavigator = ({ isAppReady }: Props) => {
   const finishLoginProgressRef = useRef<Function | null>(null);
   const waitForLoginProcessRef = useRef<Promise<boolean> | null>(null);
   const isPreventDeepLinkRef = useRef(isEmptyAccounts || !hasMasterPassword || hasConfirmations);
-  const { appPopupMap, popupHistoryMap, checkPopupVisibleByFrequency, handleButtonPress, checkPopupExistTime } =
-    useContext(AppOnlineContentContext);
-  const globalAppModalContext = useContext(GlobalModalContext);
+  const { showAppPopup } = useContext(AppOnlineContentContext);
 
   useEffect(() => {
     if (!isLocked && finishLoginProgressRef.current) {
@@ -430,65 +410,14 @@ const AppNavigator = ({ isAppReady }: Props) => {
     console.warn('AppNavigator.tsx / Error boundary: ', error, stackTrace);
   };
 
-  const getDetailCurrentRoute = (state: NavigationState | undefined) => {
-    const _currentRoute = state?.routes[state?.index];
-    if (_currentRoute) {
-      switch (_currentRoute.name) {
-        case 'Home':
-        case 'Drawer':
-          const currentHomeState = _currentRoute.state;
-          // @ts-ignore
-          const currentHomeRouteMap = currentHomeState?.routes[currentHomeState?.index];
-          const currentHomeTabState = currentHomeRouteMap?.state;
-          return currentHomeTabState?.routes[currentHomeTabState?.index];
-        default:
-          return _currentRoute;
-      }
-    }
-  };
-
   const onUpdateRoute = useCallback(
     (state: NavigationState | undefined) => {
       const _currentRoute = state?.routes[state?.index];
-      const currentDetailRoute = getDetailCurrentRoute(state);
-      const currentTransformRoute = getAppTransformRouteName(currentDetailRoute?.name) || '';
-      const currentPopupList = appPopupMap[currentTransformRoute];
-      if (currentPopupList && currentPopupList.length) {
-        const filteredPopupList = currentPopupList.filter(item => {
-          const popupHistory = popupHistoryMap[`${item.position}-${item.id}`];
-          if (popupHistory) {
-            return (
-              checkPopupVisibleByFrequency(item.repeat, popupHistory.lastShowTime, popupHistory.showTimes) &&
-              checkPopupExistTime(item.info)
-            );
-          } else {
-            return checkPopupExistTime(item.info);
-          }
-        });
-        filteredPopupList &&
-          filteredPopupList.length &&
-          globalAppModalContext.setGlobalModal({
-            type: 'popup',
-            visible: true,
-            title: filteredPopupList[0].info.name,
-            message: filteredPopupList[0].content || '',
-            buttons: filteredPopupList[0].buttons,
-            onPressBtn: url => {
-              handleButtonPress(`${filteredPopupList[0].position}-${filteredPopupList[0].id}`)('popup', url);
-            },
-          });
-      }
+      showAppPopup(_currentRoute);
       updateCurrentRoute(_currentRoute);
       setCurrentRoute(_currentRoute);
     },
-    [
-      appPopupMap,
-      globalAppModalContext,
-      popupHistoryMap,
-      checkPopupVisibleByFrequency,
-      checkPopupExistTime,
-      handleButtonPress,
-    ],
+    [showAppPopup],
   );
 
   useEffect(() => {
